Move Toolbar into Managers namespace
