+++ Assets/Scripts/Managers/DialogBoxManager.cs
@@ -1,6 +1,5 @@
 using System;
 using PAC.Colour.Compositing;
-using PAC.Drawing;
 using PAC.Extensions.UnityEngine;
 using PAC.ImageEditing;
