+++ Assets/Scripts/Managers/Finder.cs
@@ -1,4 +1,3 @@
-using PAC.Drawing;
 using PAC.Input;
 using PAC.UI.Components.Specialised.ColourPicker;
